Metadata Refreshes After Login
After login, metadata was not refreshing properly and updating
navigation. I removed the global Models for CurrentLayout and
CurrentModel, and opted for a property on the Global Layouts and
Modules collections. This allows for better tracking of what is
current, so we can refresh metadata, and still retain what module a
user was on. Also Metadata class was not blocking metadata from modules
and layouts that did not require Logins after login, due to bad logic.