Seeder Finished, Installer improved, ManyMany Relationships handled
Tweaked the abstract Module class and added in functionality for
handling of Many to Many relationships with ease now. Added Logins to
Seeder, and added all relationships and more seed data for testing.
Tweaked installer, to not be so stringent on what is getting installed
(goes with the  changes to the abstract Module object). The rest is
history. Things are going to break, since I removed some old code in
some of the Module objects, but ill fix them as I continue forward.