Backbone JS Refactor
Refactored for easier/better handling of Layouts and Layout changes.
Layouts are broken currently, however the structure is there, so its
mainly re-aligning models to views and such so that they are back in
full working order. RelateField and DepedentRelateField have been added
and are working. Need to build out backend API a bit more to handle
getting related modules and such.